Admins cannot delete themselves

+++ src/Controller/UserController.php
@@ -4,6 +4,7 @@
 namespace QuizApp\Controller;
 
 use Framework\Contracts\RendererInterface;
+use Framework\Contracts\SessionInterface;
 use Framework\Controller\AbstractController;
 use Framework\Http\Request;
 use Framework\Http\Response;
@@ -37,12 +38,18 @@ class UserController extends AbstractController
      */
     private $resultsPerPage;
 
+    /**
+     * @var SessionInterface
+     */
+    private $session;
+
     public function __construct
     (
         RendererInterface $renderer,
         RepositoryManagerInterface $repositoryManager,
         UserService $userService,
         AuthenticationService $authenticationService,
+        SessionInterface $session,
         int $resultsPerPage
     )
     {
@@ -50,6 +57,7 @@ class UserController extends AbstractController
         $this->repositoryManager = $repositoryManager;
         $this->userService = $userService;
         $this->authenticationService = $authenticationService;
+        $this->session = $session;
         $this->resultsPerPage = $resultsPerPage;
     }
 
@@ -140,6 +148,10 @@ class UserController extends AbstractController
     public function deleteUser(Request $request, array $requestAttributes): Response
     {
         $id = $requestAttributes['id'];
+        if ($id === $this->session->get('id')) {
+            //TODO Make it prettier when introducing validations
+            throw new \Exception("Cannot delete yourself");
+        }
         $this->userService->delete($id);
 
         return $this->createRedirectResponse('/listUsers');
@@ -158,9 +170,17 @@ class UserController extends AbstractController
         $criteria = $this->getCriteriaFromRequest($requestAttributes, $this->resultsPerPage);
         $userSearchResult = $this->userService->getUsers($criteria);
         $paginator = new Paginator($userSearchResult->getCount(), $currentPage, $this->resultsPerPage);
+        $sessionUserId = $this->session->get('id');
 
-        return $this->renderer->renderView('admin-users-listing.phtml',
-            ['users' => $userSearchResult->getItems(), 'paginator' => $paginator, 'order' => $requestAttributes['order']]);
+        return $this->renderer->renderView
+        (
+            'admin-users-listing.phtml',
+            [
+                'users' => $userSearchResult->getItems(),
+                'paginator' => $paginator,
+                'sessionUserId' => $sessionUserId
+            ]
+        );
     }
 
     /**
